+++ app/src/hooks/useDeleteTask.ts
@@ -7,6 +7,7 @@ import useEditTaskModal from "@hooks/useEditTaskModal";
 import useCompletedTasks from "@hooks/useCompletedTasks";
 import { deleteTask } from "@utilities/server";
 import useCurrentListId from "@hooks/useCurrentListId";
+import { vibrate } from "@utilities/haptics";
 
 function useDeleteTask() {
   const { closeModal } = useEditTaskModal();
@@ -14,6 +15,7 @@ function useDeleteTask() {
   const [isCompletedTasksIncluded] = useCompletedTasks();
   return useCallback(
     async (taskId: string) => {
+      vibrate("error");
       await deleteTask(taskId);
       closeModal();
       // Update current task list